⏰ Mejora extra – Sistema de recordatorios periódicos con alertas automáticas
