+++ src/gestores/GestorReportes.java
@@ -1,11 +1,9 @@
 package gestores;
 
+import modelo.CategoriaRecurso;
+import modelo.Prestamo;
 import modelo.RecursoBase;
 import modelo.Usuario;
-import modelo.Prestamo;
-
-import java.io.FileWriter;
-import java.io.IOException;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -17,20 +15,14 @@ public class GestorReportes {
         this.prestamos = prestamos;
     }
 
-    public List<Map.Entry<RecursoBase, Long>> recursosMasPrestados() {
+    public Map<RecursoBase, Long> recursosMasPrestados() {
         return prestamos.stream()
-                .collect(Collectors.groupingBy(Prestamo::getRecurso, Collectors.counting()))
-                .entrySet().stream()
-                .sorted(Map.Entry.<RecursoBase, Long>comparingByValue().reversed())
-                .collect(Collectors.toList());
+                .collect(Collectors.groupingBy(Prestamo::getRecurso, Collectors.counting()));
     }
 
-    public List<Map.Entry<Usuario, Long>> usuariosMasActivos() {
+    public Map<Usuario, Long> usuariosMasActivos() {
         return prestamos.stream()
-                .collect(Collectors.groupingBy(Prestamo::getUsuario, Collectors.counting()))
-                .entrySet().stream()
-                .sorted(Map.Entry.<Usuario, Long>comparingByValue().reversed())
-                .collect(Collectors.toList());
+                .collect(Collectors.groupingBy(Prestamo::getUsuario, Collectors.counting()));
     }
 
     public Map<String, Long> prestamosPorCategoria() {
@@ -38,40 +30,39 @@ public class GestorReportes {
                 .collect(Collectors.groupingBy(p -> p.getRecurso().getCategoria().name(), Collectors.counting()));
     }
 
-
     public void mostrarReporteRecursosMasPrestados() {
         System.out.println("\n📊 Recursos más prestados:");
         System.out.printf("%-30s | %s\n", "Título", "Préstamos");
         System.out.println("-----------------------------------------------");
-        recursosMasPrestados().forEach(entry ->
-                System.out.printf("%-30s | %d\n", entry.getKey().getTitulo(), entry.getValue()));
+        recursosMasPrestados().forEach((recurso, cantidad) ->
+                System.out.printf("%-30s | %d\n", recurso.getTitulo(), cantidad));
     }
 
     public void mostrarReporteUsuariosMasActivos() {
         System.out.println("\n👤 Usuarios más activos:");
         System.out.printf("%-25s | %s\n", "Nombre", "Cantidad");
         System.out.println("------------------------------------------");
-        usuariosMasActivos().forEach(entry ->
-                System.out.printf("%-25s | %d\n", entry.getKey().getNombre(), entry.getValue()));
+        usuariosMasActivos().forEach((usuario, cantidad) ->
+                System.out.printf("%-25s | %d\n", usuario.getNombre(), cantidad));
     }
 
     public void mostrarReportePorTipo() {
         System.out.println("\n📚 Préstamos por tipo de recurso:");
-        System.out.printf("%-15s | %s\n", "Tipo", "Cantidad");
+        System.out.printf("%-15s | %s\n", "Categoría", "Cantidad");
         System.out.println("-------------------------------");
-        prestamosPorTipo().forEach((tipo, cantidad) ->
+        prestamosPorCategoria().forEach((tipo, cantidad) ->
                 System.out.printf("%-15s | %d\n", tipo, cantidad));
     }
 
     public void exportarRecursosMasPrestados(String nombreArchivo) {
-        try (FileWriter writer = new FileWriter(nombreArchivo)) {
+        try (java.io.FileWriter writer = new java.io.FileWriter(nombreArchivo)) {
             writer.write("Reporte de Recursos Más Prestados\n");
             writer.write("-------------------------------\n");
-            for (var entry : recursosMasPrestados()) {
+            for (var entry : recursosMasPrestados().entrySet()) {
                 writer.write(entry.getKey().getTitulo() + " → " + entry.getValue() + " préstamos\n");
             }
             System.out.println("📁 Reporte exportado exitosamente a " + nombreArchivo);
-        } catch (IOException e) {
+        } catch (Exception e) {
             System.out.println("❌ Error al exportar el reporte: " + e.getMessage());
         }
     }
