+++ src/modelo/Audiolibro.java
@@ -1,13 +1,12 @@
 package modelo;
 
 public class Audiolibro extends RecursoBase {
-
     public Audiolibro(String titulo, String id) {
         super(id, titulo);
     }
 
     @Override
-    public TipoRecurso getTipo() {
-        return TipoRecurso.AUDIOLIBRO;
+    public CategoriaRecurso getCategoria() {
+        return CategoriaRecurso.AUDIOLIBRO;
     }
 }
