+++ src/modelo/Libro.java
@@ -1,13 +1,12 @@
 package modelo;
 
 public class Libro extends RecursoBase {
-
     public Libro(String titulo, String id) {
         super(id, titulo);
     }
 
     @Override
-    public TipoRecurso getTipo() {
-        return TipoRecurso.LIBRO;
+    public CategoriaRecurso getCategoria() {
+        return CategoriaRecurso.LIBRO;
     }
 }
