+++ src/modelo/Revista.java
@@ -1,13 +1,12 @@
 package modelo;
 
 public class Revista extends RecursoBase {
-
     public Revista(String titulo, String id) {
         super(id, titulo);
     }
 
     @Override
-    public TipoRecurso getTipo() {
-        return TipoRecurso.REVISTA;
+    public CategoriaRecurso getCategoria() {
+        return CategoriaRecurso.REVISTA;
     }
 }
